refactoring code in api (request function), logic, style, folders structure + typings, index.ts
