+++ package.json
@@ -13,7 +13,6 @@
     "@types/react-dom": "17.0.0",
     "@types/react-transition-group": "4.4.1",
     "axios": "0.21.1",
-    "bootstrap": "5.0.0-beta2",
     "react": "17.0.1",
     "react-dom": "17.0.1",
     "react-responsive-carousel": "3.2.12",
@@ -45,8 +44,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "devDependencies": {
-    "@types/bootstrap": "5.0.8"
   }
 }
