+++ public/index.html
@@ -37,9 +37,9 @@
   To begin the development, run `npm start` or `yarn start`.
   To create a production bundle, use `npm run build` or `yarn build`.
 -->
-<!--<div data-city="minsk" data-app="weather-widget"></div>-->
-<!--<div data-city="minsk" data-country="belarus" data-app="news-widget"></div>-->
-<div data-city="" data-country="RUSSIA" data-app="notes-widget"></div>
+<div data-city="minsk" data-app="weather-widget"></div>
+<div data-city="" data-country="belarus" data-app="news-widget"></div>
+<div data-city="minsk" data-country="russia" data-app="notes-widget"></div>
 
 </body>
 </html>
