+++ src/api/index.ts
@@ -1,0 +1,2 @@
+export {getNewsCity, getNewsCountry} from './newsApi';
+export {getWeather} from './weatherApi';
