+++ src/api/newsApi/index.ts
@@ -1,0 +1,1 @@
+export {getNewsCity, getNewsCountry} from './newsAPI';
