+++ src/api/weatherApi/index.ts
@@ -1,0 +1,1 @@
+export {getWeather} from './weatherAPI';
