+++ src/api/weatherApi/weatherAPI.ts
@@ -1,0 +1,15 @@
+import axios from 'axios';
+import {WeatherAPIType} from './typings';
+
+const weatherApi = axios.create({
+  baseURL: process.env.REACT_APP_WEATHER_BASE_URL
+});
+
+const apikey = process.env.REACT_APP_WEATHER_API_KEY as string;
+
+const getQuery = (city: string, apikey: string) => {
+  return `forecast?q=${city}&units=metric&appid=${apikey}`;
+};
+
+export const getWeather = (city: string) => weatherApi
+    .get<WeatherAPIType>(getQuery(city, apikey));
