+++ src/widgets/news/typings/index.ts
@@ -1,0 +1,4 @@
+export type PropsType = {
+  city?: string;
+  country?: string;
+};
