+++ src/widgets/weather/components/informationOfCity/index.ts
@@ -1,0 +1,1 @@
+export {InformationOfCity} from './InformationOfCity';
