+++ src/widgets/weather/components/informationOfCity/typings/index.ts
@@ -1,0 +1,5 @@
+import {WeatherAPIType} from 'src/api/weatherApi/typings';
+
+export type PropsType = {
+  state: WeatherAPIType | null;
+};
