+++ src/widgets/weather/typings/index.ts
@@ -1,0 +1,3 @@
+export type PropsType = {
+  city?: string;
+};
